experiments/message2expense/pyproject.toml
[project]
name = "message2expense"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.12"
dependencies = [
    "accelerate>=1.10.0",
    "blobfile>=3.0.0",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.30",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "protobuf>=4.25.0",
    "pyautogui>=0.9.54",
    "pydantic>=2",
    "pyperclip>=1.9.0",
    "sentencepiece>=0.2.1",
    "tiktoken>=0.7.0",
    "timm>=1.0.19",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.45.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "ruff>=0.12.9",
]
